Add files via upload
File Name: de_rotateHW.m

Description: This MATLAB script, authored by Marcos Faundez-Zanuy on July 7th, 2024, is designed to de-rotate or deskew handwritten samples. Deskewing is the process of correcting the tilt or rotation of handwritten text or an image to align the text lines horizontally, ensuring that the text is properly aligned for further processing or recognition. The script handles both translation normalization (centering) and rotation normalization to correct the alignment of the handwritten sample.